Update liri-extra/* with changes to DEPENDs*, versioning, git-r3, subslots

+++ liri-extra/music/music-1.0.0_p9999.ebuild
@@ -3,7 +3,7 @@
 
 EAPI=7
 
-inherit cmake-utils eutils xdg
+inherit cmake-utils eutils git-r3 xdg
 
 DESCRIPTION="A modern music app for modern people."
 HOMEPAGE="https://github.com/lirios/music"
@@ -13,11 +13,12 @@ LICENSE="GPL-3+ BSD FDL-1.3+ MPL-2.0"
 
 # Live/snapshots do not get KEYWORDS.
 
-SLOT="0/${PV}"
+SLOT="0/$(ver_cut 1-3 ${PV})"
 QT_MIN_PV=5.14
 IUSE+=" doc +system-fluid test"
 
-FLUID_DEPEND=" >=dev-libs/wayland-1.15
+FLUID_DEPEND="
+	>=dev-libs/wayland-1.15
 	>=dev-qt/qdoc-${QT_MIN_PV}:5=
 	>=dev-qt/qtcore-${QT_MIN_PV}:5=
 	>=dev-qt/qtdeclarative-${QT_MIN_PV}:5=
@@ -26,41 +27,37 @@ FLUID_DEPEND=" >=dev-libs/wayland-1.15
 	>=dev-qt/qtquickcontrols2-${QT_MIN_PV}:5=
 	>=dev-qt/qtsvg-${QT_MIN_PV}:5=
 	>=dev-qt/qtwayland-${QT_MIN_PV}:5=
-	  liri-base/qtaccountsservice"
-FLUID_BDEPEND=" >=dev-util/cmake-3.10.0
-	  virtual/pkgconfig
+	 ~liri-base/qtaccountsservice-1.3.0_p9999"
+FLUID_BDEPEND="
+	>=dev-util/cmake-3.10.0
 	>=kde-frameworks/extra-cmake-modules-1.7.0
-	>=liri-base/cmake-shared-1.0.0:0/1.1.0
-	test? ( >=dev-qt/qttest-${QT_MIN_PV}:5= )"
-DEPEND+=" >=dev-qt/qtcore-${QT_MIN_PV}:5=
+	 ~liri-base/cmake-shared-2.0.0_p9999
+	  virtual/pkgconfig
+	  test? ( >=dev-qt/qttest-${QT_MIN_PV}:5= )"
+DEPEND+="
+	>=dev-qt/qtcore-${QT_MIN_PV}:5=
 	>=dev-qt/qtdeclarative-${QT_MIN_PV}:5=
 	>=dev-qt/qtgui-${QT_MIN_PV}:5=
 	>=dev-qt/qtmultimedia-${QT_MIN_PV}:5=[qml]
 	>=dev-qt/qtquickcontrols2-${QT_MIN_PV}:5=
 	>=dev-qt/qtsvg-${QT_MIN_PV}:5=
 	>=dev-qt/qtsql-${QT_MIN_PV}:5=
-	media-libs/taglib:taglib2-preview
-	 system-fluid? ( >=liri-base/fluid-1.2.0 )
-	!system-fluid? ( ${FLUID_DEPEND} )"
+	  media-libs/taglib:taglib2-preview
+	 !system-fluid? ( ${FLUID_DEPEND} )
+	  system-fluid? ( ~liri-base/fluid-1.2.0_p9999 )"
 RDEPEND+=" ${DEPEND}"
-BDEPEND+=" >=dev-util/cmake-3.10.0
+BDEPEND+="
+	>=dev-util/cmake-3.10.0
 	  virtual/pkgconfig
-	  system-fluid? ( >=liri-base/cmake-shared-1.0.0:0/1.1.0 )
-	 !system-fluid? ( ${FLUID_BDEPEND} )"
-EGIT_COMMIT="626206c4b9f3cbd85dd5c5dc27dffb3d4098fb5d"
-FLUID_COMMIT="a99dc04a067abeb9110664f5427e9a701e6744d4"
-CMAKE_SHARED_COMMIT="4fdf2cb6edd23aaa467cfdbc758ee8ed15cbfc81"
-QBS_SHARED_COMMIT="ebee5a8798ab0a063b56cf7c73be2a28ec353a2e"
-SRC_URI="
-https://github.com/lirios/music/archive/${EGIT_COMMIT}.tar.gz
-	-> ${CATEGORY}-${PN}-${PV}-${EGIT_COMMIT:0:7}.tar.gz
-https://github.com/lirios/fluid/archive/${FLUID_COMMIT}.tar.gz
-	-> ${CATEGORY}-${PN}-fluid-${FLUID_COMMIT:0:7}.tar.gz
-https://github.com/lirios/cmake-shared/archive/${CMAKE_SHARED_COMMIT}.tar.gz
-	-> ${CATEGORY}-${PN}-cmake-shared-${CMAKE_SHARED_COMMIT:0:7}.tar.gz
-https://github.com/lirios/qbs-shared/archive/${QBS_SHARED_COMMIT}.tar.gz
-	-> ${CATEGORY}-${PN}-qbs-shared-${QBS_SHARED_COMMIT:0:7}.tar.gz"
-S="${WORKDIR}/${PN}-${EGIT_COMMIT}"
+	 !system-fluid? ( ${FLUID_BDEPEND} )
+	  system-fluid? ( ~liri-base/cmake-shared-2.0.0_p9999 )"
+SRC_URI=""
+EGIT_BRANCH="master"
+EGIT_REPO_URI="https://github.com/lirios/${PN}.git"
+EGIT_OVERRIDE_REPO_LIRIOS_QBS_SHARED="https://github.com/lirios/qbs-shared.git"
+EGIT_OVERRIDE_REPO_LIRIOS_CMAKE_SHARED="https://github.com/lirios/cmake-shared.git"
+S="${WORKDIR}/${P}"
+PROPERTIES="live"
 RESTRICT="mirror"
 _PATCHES=(
 	"${FILESDIR}/music-1.0.0_pre20200314-reference-taglib2.patch"
@@ -140,8 +137,25 @@ pkg_setup() {
 }
 
 src_unpack() {
-	unpack ${A}
-	if ! use system-fluid ; then
+	git-r3_fetch
+	git-r3_checkout
+	local v_live=$(grep -r -e "VERSION \"" "${S}/CMakeLists.txt" | head -n 1 | cut -f 2 -d "\"")
+	local v_expected=$(ver_cut 1-3 ${PV})
+	if ver_test ${v_expected} -ne ${v_live} ; then
+		eerror
+		eerror "Version bump required."
+		eerror
+		eerror "v_expected=${v_expected}"
+		eerror "v_live=${v_live}"
+		eerror
+		die
+	else
+		einfo
+		einfo "v_expected=${v_expected}"
+		einfo "v_live=${v_live}"
+		einfo
+	fi
+	if ! use system-fluid && false ; then
 		rm -rf "${S}/fluid" || die
 		ln -s "${WORKDIR}/fluid-${FLUID_COMMIT}" "${S}/fluid" || die
 		rm -rf "${S}/fluid/qbs/shared" || die
